requirements.txt
#python 3.9

black==22.3.0
click==8.1.3
Pillow==8.1.2
